# src/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/usuarios": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/lista_usuario": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id_usuario",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/vehiculos": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id_usuario",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/infoVehiculo": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id_matricula",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Servicios": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id_matricula",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/infoServicios": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id_servicio",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/editUsuario": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "id_usuario",
            "in": "query",
            "type": "string"
          },
          {
            "name": "nombre",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/crearUsuario": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "id_usuario",
            "in": "query",
            "type": "string"
          },
          {
            "name": "nombre",
            "in": "query",
            "type": "string"
          },
          {
            "name": "apellidos",
            "in": "query",
            "type": "string"
          },
          {
            "name": "dni",
            "in": "query",
            "type": "string"
          },
          {
            "name": "telefono",
            "in": "query",
            "type": "string"
          },
          {
            "name": "email",
            "in": "query",
            "type": "string"
          },
          {
            "name": "contrasena",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/eliminarUsuario": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "id_usuario",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  }
}